Moved Sally7MemoryPool to Internal-namespace
Cf. https://github.com/mycroes/Sally7/pull/23#discussion_r728385327
